Games.Platform: Pull in libraries whose development files are in the Sdk
Development files for libgles2, libglu1-mesa are indirectly pulled in
by development files for SDL2. Development files for libxcb-randr0,
libxcb-render0 are indirectly pulled in by development files for
Mesa, via libxcb-present-dev.

It's misleading if we have development files for something in the Sdk
(so we can link against it at build time) but then don't ship the
corresponding shared library in the Platform.