Builder: Allow removing a suffix from source versions
This is necessary for packages that were built in OBS, which pretend
to be a sourceful upload to avoid breaking packages' assumptions that
Architecture: all packages are never binNMU'd and hence can be depended
on as foo-data (= ${source:Version}). However, the source isn't kept,
and is identical to the base version except for debian/changelog.

Signed-off-by: Simon McVittie <smcv@collabora.com>