Try harder to get packages' source code
In recent versions of apt, "apt-get source PACKAGE=VERSION" implies
APT::Get::Only-Source=true (which interprets PACKAGE as a source
package and not a binary package), but Ubuntu precise's apt doesn't
seem to imply that, and for some reason downloading everything in one
big batch doesn't seem to give the same answer as doing individual
sources. Try to do one big batch because it's faster and has better
progress feedback, but be willing to fall back if apt gets confused.

Signed-off-by: Simon McVittie <smcv@collabora.com>