heavy SDK: Disable setuid dbus-daemon-launch-helper
After patching xterm to disable libutempter, this becomes the only file
in the SDK sysroot owned by a user or group outside the range statically
allocated by Debian Policy (0-99). We don't need system bus activation
for a chroot or container, and systems like schroot assume that
/etc/passwd and /etc/group can safely be shared between the host and
the chroot, so defang it by making it root:root 0444.

The dpkg-statoverride record remains in place, so if you are using this
SDK to make a complete virtual machine for some reason, you can use
`apt-get --reinstall install dbus` to reverse this.

Signed-off-by: Simon McVittie <smcv@collabora.com>