set-build-id: Populate /etc/debian_chroot
This file is read by the default /etc/bash.bashrc in Debian and
included as a prefix to the prompt, in this style:

    (contents of /etc/debian_chroot) user@host:~$

It is automatically populated with the chroot name by schroot, but not
by other container tools like Docker and Flatpak. We might as well
give it a useful value for all Flatdeb-generated containers.

Signed-off-by: Simon McVittie <smcv@collabora.com>