Sdk: Create build-ID-based hard links to detached debug symbols
Legacy path-based names are more difficult to use against a sysroot:
if your sysroot is in ~/sysroot and your detached debug symbols are
in ~/symbols, and you want to find detached debug symbols for
~/sysroot/usr/lib/x86_64-linux-gnu/libfoo.so.0 by using
"gdb -iex set-debug-file-directory /home/me/symbols:/usr/lib/debug",
then you would have to create a symlink
~/symbols/home/me/sysroot/usr pointing to ~/symbols/usr. Build-ID-based
names do not have this issue.

This change is based on code in steam-runtime, which uses symbolic
links. Hard links are a somewhat better fit for OSTree, since they
are automatically deduplicated, and also have the advantage that
multiple .build-id directories can easily be combined with rsync.

Signed-off-by: Simon McVittie <smcv@collabora.com>