Ensure that /usr/lib32/locale and /usr/lib64/locale exist
Most Linux distributions either use Red-Hat-style biarch (lib and lib64)
or Debian-style multiarch (lib/i386-linux-gnu and lib/x86_64-linux-gnu),
and configure glibc to look for locales in /usr/lib/locale. glibc itself
has special cases to do this for common "lib64" architectures like
x86_64. This is analogous to the way systemd user units and tmpfiles are
always in /usr/lib/systemd, never in /usr/lib64/systemd or
/usr/lib/*-linux-gnu/systemd.

However, Arch Linux and its derivatives use currently a variation of
biarch with lib32 and a 64-bit lib, and create a symlink
/usr/lib32/locale -> ../lib/locale rather than configuring the 32-bit
glibc to look in /usr/lib/locale in the first place.

When using pressure-vessel we might have to use a host system glibc,
so we need to make locales available in both locations. Create the same
symlink that Arch Linux does. For symmetry, also create one in /usr/lib64
when we build 64-bit-capable runtimes.

Signed-off-by: Simon McVittie <smcv@collabora.com>