collect-dbgsym: Install all -dbgsym packages in one transaction
If we have (for example) 400 -dbgsym packages to install, invoking apt
400 times is rather slow due to startup overhead (locking, reading
and updating the database). Instead, invoke apt-cache once to find out
what is available, and use that to compose a single very large apt-get
command-line.

Similarly, we can summarize what was and wasn't installed from the
information we already have, without having to invoke apt a lot.

Signed-off-by: Simon McVittie <smcv@collabora.com>