platformize: Fix systemd-nspawn error when running in gitlab CI
When running in gitlab CI with docker the following error may occur:
platformize | Failed to register machine: Unit machine-root.scope already exists.

We only use systemd-nspawn for temporary containers which are not intended
to be used interactively, so it is not useful for them to be listed by
machined.

Registering with machined can also be harmful: because our root directory
is named `root`, if two concurrent flatdeb instances communicate with the
same machined instance, the second one will try to use the same name as
the first and so will fail to register, causing flatdeb to fail.

This can be avoided by preventing the systemd-nspawn container from
registering with machined (which is supported since systemd v209).

Signed-off-by: Frédéric Danis <frederic.danis@collabora.com>