apps: Add strip_source_version_suffix support
This is necessary for packages that were built in OBS which adds a suffix,
but the source isn't kept, and is identical to the base version except
for debian/changelog.
This is already done for runtimes.
The same `strip_source_version_suffix` variable is used for both runtimes
and apps.

Signed-off-by: Frédéric Danis <frederic.danis@collabora.com>