Add support to install additional packages to the base chroot
This gives us a way to drop in configuration changes (in particular apt
pinning preferences) as soon as possible after the debootstrap run,
allowing non-default pinning to be used to install the extra packages
for the Platform and SDK.

Please note that `$keyringpackage`, and the automatic installation of
`apt-transport-https`, have been removed because this new option can be
considered a generalization of both of them.
So if your suite needs `apt-transport-https`, or if you were using
`$keyringpackage`, list the needed packages with the
`additional_base_packages` option instead.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>