runtimes: Make Steam compatibility tools available
As requested and later documented in
<https://github.com/ValveSoftware/steam-for-linux/issues/6310>,
Steam looks for compatibility tools in
/usr{/local,}/share/steam/compatibilitytools.d. This is problematic
for pressure-vessel, because pressure-vessel reserves /usr for the
runtime and mounts the host's /usr in /run/host/usr, but Steam will
pass an absolute path in /usr to be used to launch the compat tool.
We can make this no longer a problem by creating symlinks to the
host locations.

Partial fix for https://github.com/ValveSoftware/steam-runtime/issues/302
(we also need to mount $STEAM_EXTRA_COMPAT_TOOLS_PATHS, but that has to
be done by changing pressure-vessel).

Signed-off-by: Simon McVittie <smcv@collabora.com>