make-flatpak-friendly: Don't remove /usr/local
When building a runtime intended to interoperate with a wider system,
it will sometimes be necessary to put symlinks in /usr/local to make
files from /run/host/usr/local available. For example, Steam looks for
compatibility tools in /usr/local/share/steam/compatibilitytools.d and
passes them to processes in the pressure-vessel container by their
absolute path.

/usr/local is also a reasonable place to put any unpackaged software
that is installed by the flatdeb configuration.

Signed-off-by: Simon McVittie <smcv@collabora.com>