runtimes: Convert hard-linked DRI drivers and interpreters to symlinks
The Steam CDN probably doesn't understand that hard links are a thing
that can exist, so will see these as multiple identical files, wasting
a lot of disk space (particularly for the DRI drivers, some of which
statically link LLVM).

I've only done this for larger files: for small files like gunzip it
isn't worth it.

Signed-off-by: Simon McVittie <smcv@collabora.com>