suites: Remove misleading comments
heavy and soldier have public apt repositories now, so the sources.list
that we were already generating will update from those repositories.

heavy follows the same pattern as scout: the beta apt suite tracks
the beta Steam client, and the stable apt suite (which is the one we
use in practice) tracks the general-availability Steam client.

soldier is similar, but because it isn't bundled in Steam, the beta and
stable apt suites track the beta and stable SteamLinuxRuntime_soldier
depots instead.

sniper and medic are not in production use and not yet available to the
public, but it is very likely that when they appear, they will have
the same setup as soldier, so the comments in those files don't
really add any value either.

Signed-off-by: Simon McVittie <smcv@collabora.com>