Consistently use archive.debian.org for Debian 8 'jessie'
The jessie suite exists on both deb.debian.org and archive.debian.org
for now, but archive.debian.org is more future-proof, so let's move
to that one now.

jessie/updates still exists in the security archive and is not on
archive.debian.org yet, so leave the security URLs unchanged.

Signed-off-by: Simon McVittie <smcv@collabora.com>