scout: force install the latest apt version
In our repositories we have a backported `apt` from Debian 8.
However, that version will be kept back by apt-get because the
dependencies changed.
Additionally, we don't currently provide the `libapt-pkg4.12-dbgsym`
(due to https://gitlab.steamos.cloud/infrastructure/tasks/-/issues/13),
and that will break the dist-upgrade with:
" libapt-pkg4.12 : Breaks: apt (< 0.9.4~) but
0.8.16~exp12ubuntu10.29+srt1 is to be installed"

For this reason we explicitly install `apt` and `libapt-pkg4.12` to grab
the latest available versions from the beginning.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>