platformize: Remove sudo
There's no point in having a privilege elevation tool in a container
that is only going to be used in a Flatpak-style environment with no
privileged uids and NO_NEW_PRIVS. However, we do want sudo in any
sysroot that will be used as a Docker/OCI container, because it's
required by Toolbx. Allow it to be installed at the Platform sysroot
stage, then removed before we build the Platform runtime.

Signed-off-by: Simon McVittie <smcv@collabora.com>